feat: Playlist의 갯수 가져오기 추가

+++ src/main/java/toolc/yourlist/playlist/domain/SaveRequest.java
@@ -10,10 +10,10 @@ public class SaveRequest {
   private final String loginId;
   private final String title;
   private final boolean isMember;
-  private final int playlistCount;
+  private final long playlistCount;
 
   @Builder
-  public SaveRequest(String loginId, String title, boolean isMember, int playlistCount) {
+  public SaveRequest(String loginId, String title, boolean isMember, long playlistCount) {
     this.loginId = loginId;
     this.title = title;
     this.isMember = isMember;
